HYD-5682 Parsely: Update to pass back error info
Parsely has been using token lengths to determine the
amount of match that occured before an error. In addition, we have
not been rewinding tokens in every parser on error. Finally, we mutate
the tokens array to reflect consumption instead of passing it back.

Update parsely to return back the tokens and the number of tokens
consumed. This should make it so we don't have to rewind tokens anymore.
In addition, we can easily tell how many tokens have been consumed so
far without diffing token arrays.

  - Create module.
  - Switch to return tokens, consumed and results from each parser.
  - Update parser combinators to expect objects of tokens, consumed and
    results being returned.

Signed-off-by: Joe Grund <joe.grund@intel.com>