HYD-5841 Parsely: create sep-by-infinity parser
The concept of having a useful autocomplete means using a parser under
the hood that is not satisfied in most cases.

For instance, if we have the expression:

a=b

A normal parser will consider that a complete expression.

However, with an autocomplete, we always want to suggest another set of
tokens, so we need to consider that expression incomplete.

Create a parser sep-by-infinity that will essentially never be
satisfied. It will return an error if either the symbol or sep parsers
do not match and will never return output.

By doing this, we ensure we alway suggest either the result of
symbolFn or sepFn in the autocomplete.

  - Create sep-by-infinity parser + tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>