HYD-5848 Parsely: create token-error parser
parsely works by using a functional concept called the either type
(although in a simplified fashion).

Basically, the output of a parser can be either the result of the
expected value or an error instance.

The token parser has a way to transform the value output if needed. It
was lacking a way to transform the error output. This is now more
important because the autocomplete will need more precise output, for
example knowing an error originated from a left-hand value.

  - Add token-error parser and add tests.
  - Reimplement token using token-error.
  - Extract result as type in token-error.

Signed-off-by: Joe Grund <joe.grund@intel.com>