HYD-5851 Parsely: choice should aggregate results by match index
The choice parser currently takes the last error with the longest match
as the result.

This is sub-optimal as we know all the missed choices, we just don't
aggregate them.

Instead we can aggregate all failed choices and roll up the results in a
reduce operation.

By doing this, we can enumerate all the errors and suggestions.

  - Update choice parser to aggregate errors by index.
  - Reformat tests for less diff churn.

Signed-off-by: Joe Grund <joe.grund@intel.com>