HYD-5885 Parsely: add match parser
We frequently find ourselves wanting to match specific values when
parsing.

This is currently done in the lexer, which defines specific value subset
tokens on top of generic value tokens.

The issue here is this approach is not flexible enough. For example a
token in may match, in or info.

There is really a question if the lexer buys us much beyond character
start / end locations. We should probably just use it for very generic
matching and leave more specific matches up to the parser layer.

In this vein, we can add a match parser. This parser will match the
content of a lexed token. In this way we can ensure full matches and not
substring matches.

  - Add match parser.
  - Add tests.
  - Update deps.

Signed-off-by: Joe Grund <joe.grund@intel.com>