HYD-5993 Create ParseError type and split token parsing
Currently we are trying to model suggestions in the result type.
We don't need to do this as we are interested on suggestions on a
parse failure.

Instead we can create a new type, ParseError that itself can hold
suggestions and start / end postions.

In addition, ParseError can have an adjust method that returns a new
ParseError, allowing consumers to adjust the expectations in the error
message.

Our token function also does input / output parsing, and we tried to do
a match function as well for more specific parsing on token types.

Instead, we can use composition to alter success and error messages in
results, which will keep the parsers themselves cleaner.

  - Remove match parser.
  - Remove chainL1 parser.
  - Implement ParseError, onSuccess, and onError.
  - Remove jscs.
  - Bump deps.
  - Add tokenTo parser.
  - Add matchValue parser.

Signed-off-by: Joe Grund <joe.grund@intel.com>