In order to handle . in values a many1 parser can be utilized. The many1 parser will function as follows:
When parsing a list of tokens:
1. If parsing the first token results in an error, the error will be
returned along with the list of tokens. Zero items will be consumed.
2. If parsing the first token does not result in an error, the parser
will continue parsing until it encounters an error on one of the tokens
in the list. The result will contain the list of un-consumed tokens, the
number of times a token was successfully consumed, and the result stream
in the form of consumed tokens appeneded together.

fixes #14

Signed-off-by: Will Johnson <william.c.johnson@intel.com>